+++ packages/website/src/pages/feedback/Loading.tsx
@@ -12,6 +12,18 @@ const fullContent = `
     export default () => <Loading fullContent />
 `;
 
+const loadingSpinner = `
+    import {LoadingSpinner} from '@coveord/plasma-react';
+
+    export default () => (
+        <>
+            <LoadingSpinner size={16} />
+            <LoadingSpinner />
+            <LoadingSpinner size={32} />
+        </>
+    );
+`;
+
 export default () => (
     <PageLayout
         id="Loading"
@@ -22,6 +34,7 @@ export default () => (
         code={code}
         examples={{
             fullContent: {code: fullContent, title: 'Vertically centered'},
+            loadingSpinner: {code: loadingSpinner, title: 'Loading spinner that can be used in other components'},
         }}
     />
 );